Add unit to TagInfo objects
When using pre-defined units on a tag, the value
returned by the TagInfo class is an enum such as 'A42'
or 'CEL'. This is a limitation and to return the
proper value, the tag must use the 'custom unit' option.

See https://techforum.ewon.biz/thread-1667.html for the
related technical forum.