Update getDayHourMinSecsForMillis to use long
This change does not affect existing implementations as Java
will natively promote int to long. Using long makes more sense
though, as time is usually expressed and stored as a long in
Java. For example, our extensions library method
HistoricalDataQueueManager#getQueueTimeBehindMillis() returns
a long value, which is commonly passed to this method.