Added Reboot Watchdog
Utility class for handling cyclic reboots of a Java application running
on the Ewon Flexy system. At startup the reboot watchdog will count the
reboot, once the application is stable and has been running for a
sufficient amount of time the application should service the watchdog.
Servicing the watchdog clears the reboot count. If the watchdog is not
serviced before another reboot the reboot count is again incremented.
If the reboot count gets too high the jvmrun file is disabled. This
functionality helps prevent cyclic reboots in faulty applications.
Devices in these reboot loops can be difficult to access via http, ftp,
or vpn.