Update IDE Files
Changes to the misc.xml file and compiler.xml file
are related to changes made in a recent update to
IntelliJ IDEA. This update improved Maven integration
but changed project configuration formatting slightly.

The project name change in compiler.xml is due to the
project name now being always imported from Maven. This
previously could be changed to a custom name.