Added rebootwatchdog service timer
Added a timer to the rebootwatchdog to ensure that the
application has been running for a specified amount of time
before allowing the watchdog to be serviced. This is to
ensure the application is in a stable state regardless of
the application implementation. A default wait time of 0
seconds is used if the application does not specify a wait
time. This preserves compatibility with existing implementations.

Closes #70