Added RawNumberValueUtilsTest.java
Added a test suite class for the RawNumberValueUtils
class, which tests each of the methods using a defined
number of test runs, each with random values.

This was implemented as a separate test class due to
the large number of methods. It has also been placed
in a package corresponding with the associated class
under test.