Added TagInfoTest.java
Added a test suite class for the TagInfo class to
better organize the basic tag info test method
which was in the root LibraryTest class.

Having a test suite class specific to the class under
test allows for easier modularization in the future, and
prevents confusion from large, messy test suite classes
with many tests that are unrelated.