Update DataQuality.java
Updated DataQuality.java to include versioning, which
allows for better identification for portions of the
class which are new or already existed.

Support for getting the String equivalent of the data
quality value has also been added. Lastly, a method was
added to allow for retrieving the corresponding quality
object from a specified OPCUA quality value.