Add DataPointNumber
Added the DataPointNumber class which provides
a generic DataPoint implementation for all
number types, including long (dword), float,
integer, etc.

This class is intended to supplement existing
functionality and provide a more flexible
interface for working with numerical values.
In the future, it could be expanded to be the
superclass of current numeric DataPoint classes.