Add InstantValuesEbdString.java
Added the new InstantValuesEbdString class which
provides a class wrapper for Ewon instant values
(IV) EBD requests.

This class also uses an improved method for parsing
the semicolon-delimited CSV format returned in the
result. If future implementations of EBD requests
are added, this functionality may be generally of
use and could be abstracted in to a utility class
if necessary.