Update HistoricalDataManager.java
Update the underlying EBD request parameters to
migrate towards usage of the ISO 8601 formatted
timestamps output directly by the Ewon. This is
not a full replacement for injecting local time
information in to the JVM, but does prevent any
possible timestamp errors caused by labeling of
data points with an incorrect time zone offset.

This also establishes a Javadoc versioning scheme
for the class. This value has been set to v3.0.0,
which accounts for major-version v1.0.0 being the
initial release, and v2.0.0 being after the rapid
catch-up functions were added.