Add ISO 8601 timestamp support in DataPoint class
The DataPoint class will now have two timestamps, one as seconds from
epoch and another as ISO 8601 string. Add a method to get the ISO 8601
timestamp. Mark getTimeStamp method as deprecated. This is because it
is never clear to the caller if the value is relative to UTC or local
time.