LocalTimeOffsetCalculator: set GMT for parsing
This change sets the simpleDateFormat (for local
time parsing) to have the GMT time zone.
Move GMT timezone name to TimeLibConstants and
make public.

This is necessary because the function finds the
difference between system time (epoch millisecond
relative to GMT) to local time ( epoch millisecond
relative to GMT). Before setting the default time
zone for the JVM, this explicit setting is not
necessary. However, after setting the default
time zone, the simpleDateFormat will use the
default time zone, which is often not GMT. This
causes the time difference to be incorrect.